requirements.txt
Pillow==9.4.0
PySnooper==1.1.1
